# docs/yggdrasil/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ratatoskr WebPanel(Yggdrasil) Rest API",
        "title": "Ratatoskr(Yggdrasil)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "ali.kaviani@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5050",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Pod/Container health check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "General"
                ],
                "summary": "Pod/Container health check",
                "operationId": "handle_health_check",
                "responses": {
                    "200": {
                        "description": "Success Result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.HealthCheckResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.HealthCheckResponse": {
            "type": "object",
            "properties": {
                "binary_uptime": {
                    "type": "string"
                },
                "build_time": {
                    "type": "string"
                },
                "container": {
                    "type": "string"
                },
                "git_commit": {
                    "type": "string"
                },
                "go_version": {
                    "type": "string"
                },
                "kernel_uptime": {
                    "type": "string"
                },
                "load_average_fifteen": {
                    "type": "number"
                },
                "load_average_five": {
                    "type": "number"
                },
                "load_average_one": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 0
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}